Migrate setup.py to pyproject.toml #11057
