+++ arches/app/utils/data_management/resources/importer.py
@@ -33,7 +33,7 @@ from arches.app.models.models import (
 from arches.app.models.system_settings import settings
 from arches.app.utils.betterJSONSerializer import JSONSerializer, JSONDeserializer
 from arches.app.utils.i18n import capitalize_region
-from arches.setup import unzip_file
+from arches.app.utils.zip import unzip_file
 from .formats.csvfile import CsvReader
 from .formats.archesfile import ArchesFileReader
 import ctypes
