--- arches/install/arches-templates/project_name/install/requirements.txt
@@ -1,1 +1,0 @@
-arches=={{ arches_semantic_version }}
