--- arches/install/arches-templates/project_name/install/requirements_dev.txt
@@ -1,6 +1,0 @@
-livereload
-sst
-coverage
-django-silk==5.1.0
-pre-commit
-black==24.4.2
