--- arches/install/requirements.txt
@@ -1,37 +1,0 @@
-Django>=4.2.13,<5.0.0
-psycopg2==2.9.9
-urllib3<2
-elasticsearch>=8.3.1,<9.0.0
-rdflib==4.2.2
-django-guardian==2.4.0
-python-memcached==1.59
-celery==5.3.6
-django-celery-results==2.5.1
-SPARQLWrapper==1.8.5
-django-recaptcha==3.0.0
-edtf==4.0.1
-django-revproxy==0.12.0
-django-cors-headers==4.2.0
-django-oauth-toolkit==1.7.0
-polib==1.1.1
-PyLD[requests]==1.0.5
-pyprind==2.11.3
-pycryptodome<4.0.0,>=3.3.1
-pyotp>=2.6.0
-qrcode>=7.3.1
-pyshp==2.1.2
-requests[security]>=2.31.0
-python-slugify==7.0.0
-pillow>=7.0.0
-arcgis2geojson==2.0.0
-openpyxl==3.1.2
-django-webpack-loader==2.0.1
-filetype==1.2.0
-defusedxml==0.7.1
-requests-oauthlib==1.3.1
-pyjwt>=2.0.0,<3
-pyjwt[crypto]
-semantic-version==2.10.0
-tzdata==2023.3
-pytz==2023.3
-django-ratelimit==4.1.0
