+++ arches/management/commands/packages.py
@@ -8,7 +8,6 @@ import sys
 import urllib.request, urllib.parse, urllib.error
 import os
 import logging
-from arches.setup import unzip_file
 from arches.management.commands import utils
 from arches.app.utils.i18n import LanguageSynchronizer
 from arches.app.utils import import_class_from_string
@@ -30,6 +29,7 @@ from arches.app.utils.data_management.resources.formats.format import (
     Reader as RelationImporter,
 )
 from arches.app.utils.data_management.resources.exporter import ResourceExporter
+from arches.app.utils.zip import unzip_file
 from arches.app.models.system_settings import settings
 from arches.app.models import models
 from arches.app.models.fields.i18n import I18n_String
