+++ releases/7.6.0.md
@@ -118,7 +118,7 @@ JavaScript:
         eslint-config-prettier == 9.1.0
         eslint-plugin-vue == 9.20.0
         prettier == 3.3.0
-        primevue == 4.0.0-beta.3
+        primevue == 4.0.0-rc.2
         nodemon == 3.0.2
         sass == 1.70.0
         ts-loader == 9.5.1
@@ -158,6 +158,8 @@ Minor incompatibilities:
 -   `BaseImportModule.stage_files` is now abstract. (It used to call
     `stage_excel_files()`, which was only defined on certain subclasses.) `stage_files()`
     is now a more attractive target for overriding than `run_load_task()`.
+-   `unzip_file()` moved from `arches.setup` to `arches.app.utils.zip`
+-   Version-related utils moved from `arches.setup` to `arches.version`
 
 
 ### Deprecations
@@ -342,14 +344,9 @@ Minor incompatibilities:
 
 12. If you do any custom development, update and reinstall the python development dependencies, and install the pre-commit hooks. (You can always ignore these hooks with `git commit --no-verify`.)
 
-    1. Add the following lines to ${project_name}/install/requirements_dev.txt
+    1. Navigate to the directory with your project's .pre-commit-config.yaml file.
         ```
-        pre-commit
-        black==24.4.2
-        ```
-    2. Navigate to the directory with your project's .pre-commit-config.yaml file.
-        ```
-        pip install -r ${project_name}/install/requirements_dev.txt
+        pip install -e '.[dev]'
         pre-commit install
         ```
