+++ tests/test_settings.py
@@ -22,7 +22,7 @@ import os
 
 try:
     from django.utils.translation import gettext_lazy as _
-except ImportError:  # unable to import prior to installing requirements.txt in setup.py
+except ImportError:  # unable to import prior to installing requirements
     pass
 
 PACKAGE_NAME = "arches"
